requirements-upload.txt
pypdf
langchain
langchain-community
langchain-core
langchain-postgres
langchain-ollama
supabase
psycopg2-binary
pgvector
python-dotenv
